% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figures.R
\name{inf.plot}
\alias{inf.plot}
\title{A function for plotting inferred temporal components (trend and seasonality) with 95\% credible intervals.}
\usage{
inf.plot(inf.object)
}
\arguments{
\item{inf.object}{Resulting fit object from "SMOOTHING_INFERENCE".}
}
\value{
Figures
}
\description{
A function for plotting inferred temporal components (trend and seasonality) with 95\% credible intervals.
}
\examples{
sim_adjmat<- matrix(0, nrow = 9, ncol = 9)
uppertriang<- c(1,0,1,0,0,0,0,0,1,0,1,0,0,0,0,0,0,1,0,0,0,1,0,1,0,0,1,0,1,0,0,0,1,1,0,1)
gdata::upperTriangle(sim_adjmat, byrow=TRUE)<- uppertriang
gdata::lowerTriangle(sim_adjmat, byrow=FALSE)<- uppertriang
set.seed(0); mod0<- simulateMultiModel(Modeltype = 0, time = 60, nstrain = 5, adj.matrix = sim_adjmat)
MCMCfit0<- SMOOTHING_INFERENCE(y=mod0[["y"]], e_it=mod0[["e_it"]], Modeltype = 0, sim_adjmat, MCMC_iterations = 15000, HMC_iterations = 5000, Stan = FALSE)
inf.plot(MCMCfit0)
}
