% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modelcomparison.R
\name{ModelEvidence}
\alias{ModelEvidence}
\title{Title}
\usage{
ModelEvidence(
  y,
  e_it,
  adjmat,
  Modeltype,
  inf.object,
  num_samples = 5000,
  y_total = NULL
)
}
\arguments{
\item{y_total}{}
}
\description{
Title
}
