% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{simulateMultiModel}
\alias{simulateMultiModel}
\title{Simulation from the models proposed in (Adeoye et al. 2026)}
\usage{
simulateMultiModel(
  Modeltype,
  time,
  nstrain,
  adj.matrix,
  copulaParam,
  e_it = matrix(c(rep(c(rpois(time, 5e+05), rpois(time, 1e+06)), 4), rpois(time, 5e+05)),
    byrow = T, ncol = time),
  B = runif(nstrain),
  T.prob = matrix(c(0.9, 0.1, 0.2, 0.8), nrow = 2, byrow = T),
  r = sim.RW2mean0(time, sd = 0.009),
  s = sim.Seasonals(Amplitude = 1.4),
  u = sim.Spatials(adj.matrix)
)
}
\arguments{
\item{Modeltype}{The model specification (0 = No epidemic model, 1 = Independent 1 model, 2 = Independent 2 model, 3 = Gaussian factor copula 1 model, 4 = Gaussian factor copula 2 model, 5 = Frank copula 1 model, 6 = Frank copula 2 model, 7 = General-dependent model).}

\item{time}{Temporal dimension}

\item{nstrain}{Number of disease types}

\item{adj.matrix}{The adjacency matrix describing the connectivity/neighborhood structure of the spatial locations.}

\item{copulaParam}{Copula parameter(s)}

\item{e_it}{A space-time matrix showing the number of susceptible individuals (locations on the row, and time on the columns).}

\item{B}{Strain-specific epidemic effects}

\item{T.prob}{Transition probabilities}

\item{r}{Trend component}

\item{s}{Seasonal component}

\item{u}{spatial component}
}
\value{
A 3D array of multi-type disease counts (dim=c(locations, time, types)).
}
\description{
Simulation from the models proposed in (Adeoye et al. 2026)
}
\examples{
set.seed(0)
sim_adjmat<- matrix(0, nrow = 9, ncol = 9)
uppertriang<- c(1,0,1,0,0,0,0,0,1,0,1,0,0,0,0,0,0,1,0,0,0,1,0,1,0,0,1,0,1,0,0,0,1,1,0,1)
gdata::upperTriangle(sim_adjmat, byrow=TRUE)<- uppertriang
gdata::lowerTriangle(sim_adjmat, byrow=FALSE)<- uppertriang
set.seed(0); mod0<- simulateMultiModel(Modeltype = 0, time = 60, nstrain = 5, adj.matrix = sim_adjmat)
}
