% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inference.R
\name{SMOOTHING_INFERENCE}
\alias{SMOOTHING_INFERENCE}
\title{Inference Method 1 -- Smoothing}
\usage{
SMOOTHING_INFERENCE(
  y,
  e_it,
  Modeltype,
  adjmat,
  step_sizes = list(r = 0.3, s = 0.3, u = 0.3),
  Stan = FALSE,
  MCMC_iterations = 15000,
  HMC_iterations = 5000,
  sdBs = 0.003,
  sdGs = 0.005,
  sdLambdas = 0.003,
  sdCop = 2e-04,
  y_total = NULL,
  RM_Gs = TRUE,
  RM_Bs = TRUE,
  RM_Cop = TRUE,
  RM_Lambdas = TRUE,
  burn_in = 3000,
  n_chains = 4,
  GPU = FALSE,
  adaptdelta = 0.9,
  verbose = TRUE,
  initial_values = NULL,
  use_initals = FALSE
)
}
\arguments{
\item{y}{A 3D array of multi-type disease counts (dim=c(locations, time, types)).}

\item{e_it}{A space-time matrix showing the number of susceptible individuals (locations on the row, and time on the columns).}

\item{Modeltype}{The model specification (0 = No epidemic model, 1 = Independent 1 model, 2 = Independent 2 model, 3 = Gaussian factor copula 1 model, 4 = Gaussian factor copula 2 model, 5 = Frank copula 1 model, 6 = Frank copula 2 model, 7 = General-dependent model).}

\item{adjmat}{The adjacency matrix describing the connectivity/neighborhood structure of the spatial locations.}

\item{step_sizes}{Step sizes for the Riemanian manifold MALA update of temporal and spatial components (Default step-sizes in the package works great).}

\item{Stan}{A logical argument asking whether to fit the model using Stan's HMC via cmdstanr.}

\item{MCMC_iterations}{The required number of iterations for inference via MCMC. Default set to 15000.}

\item{HMC_iterations}{The required number of iterations for inference via HMC. Default set to 5000.}

\item{sdBs}{Standard deviation for updating strain-specific regression parameters}

\item{sdGs}{Standard deviation for updating transition parameters}

\item{sdLambdas}{Standard deviation for updating factor loadings}

\item{sdCop}{Standard deviation for updating Frank copula parameter}

\item{y_total}{A space-time matrix showing the total untyped disease counts if available (locations on the row, and time on the columns).}

\item{RM_Gs}{A logical argument asking whether to apply Robbin-Monro tuning to sdGs}

\item{RM_Bs}{A logical argument asking whether to apply Robbin-Monro tuning to sdBs}

\item{RM_Cop}{A logical argument asking whether to apply Robbin-Monro tuning to sdCop}

\item{RM_Lambdas}{A logical argument asking whether to apply Robbin-Monro tuning to sdLambdas}

\item{burn_in}{MCMC burn-in period. Default set to 3000.}

\item{n_chains}{The number of parallel MCMC chains to run using Stan's HMC.}

\item{GPU}{A logical argument for accelerating Stan's computations using OpenCL (works with GPU only).}

\item{adaptdelta}{Target average proposal acceptance probability during Stan's adaptation period. Default set to 0.90.}

\item{verbose}{A logical argument asking whether to display fitting progress and/or error messages.}

\item{initial_values}{(Optional) specify initial values for the chain}

\item{use_initals}{(Optional)  A logical argument to use specified initial_values}
}
\value{
Samples from the target posterior distribution.
}
\description{
Inference Method 1 -- Smoothing
}
\examples{
set.seed(0)
sim_adjmat<- matrix(0, nrow = 9, ncol = 9)
uppertriang<- c(1,0,1,0,0,0,0,0,1,0,1,0,0,0,0,0,0,1,0,0,0,1,0,1,0,0,1,0,1,0,0,0,1,1,0,1)
gdata::upperTriangle(sim_adjmat, byrow=TRUE)<- uppertriang
gdata::lowerTriangle(sim_adjmat, byrow=FALSE)<- uppertriang
set.seed(0); mod0<- simulateMultiModel(Modeltype = 0, time = 60, nstrain = 5, adj.matrix = sim_adjmat)
MCMCfit0<- SMOOTHING_INFERENCE(y=mod0[["y"]], e_it=mod0[["e_it"]], Modeltype = 0, sim_adjmat, MCMC_iterations = 15000, HMC_iterations = 5000, Stan = FALSE)
mcmc.plot(MCMCfit0)
inf.plot(MCMCfit0)

}
