% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modelcomparison.R
\name{ModelEvidenceBridgeSamplingPackage}
\alias{ModelEvidenceBridgeSamplingPackage}
\title{Title}
\usage{
ModelEvidenceBridgeSamplingPackage(
  y,
  e_it,
  adjmat,
  Modeltype,
  inf.object,
  y_total = NULL,
  cores = 1
)
}
\arguments{
\item{cores}{}
}
\description{
Title
}
