% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figures.R
\name{multitypeFig}
\alias{multitypeFig}
\title{Title}
\usage{
multitypeFig(y, Modeltype = "")
}
\arguments{
\item{y}{Multitype data}

\item{Modeltype}{Model name (optional)}
}
\value{
Plot of the multitype data
}
\description{
Title
}
\examples{
set.seed(0)
mod0<- simulateMultiModel(Modeltype = 0, time = 60, nstrain = 5, adj.matrix = sim_adjmat)
multitypeFig(y=mod0[["y"]])
}
