% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Posteriormultstrain.Decoding}
\alias{Posteriormultstrain.Decoding}
\title{Title}
\usage{
Posteriormultstrain.Decoding(
  y,
  e_it,
  inf.object,
  Modeltype,
  y_total = NULL,
  thinningL = 1000,
  burn.in = 1000
)
}
\arguments{
\item{burn.in}{}
}
\description{
Title
}
