% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{simulateMultiModel}
\alias{simulateMultiModel}
\title{Title}
\usage{
simulateMultiModel(
  Modeltype,
  time,
  nstrain,
  adj.matrix,
  copulaParam,
  e_it = matrix(c(rep(c(rpois(time, 5e+05), rpois(time, 1e+06)), 4), rpois(time, 5e+05)),
    byrow = T, ncol = time),
  B = runif(nstrain),
  T.prob = matrix(c(0.9, 0.1, 0.2, 0.8), nrow = 2, byrow = T),
  r = sim.RW2mean0(time, sd = 0.009),
  s = sim.Seasonals(Amplitude = 1.4),
  u = sim.Spatials(adj.matrix)
)
}
\arguments{
\item{Modeltype}{Model specification}

\item{time}{Temporal dimension}

\item{nstrain}{Number of disease types}

\item{adj.matrix}{Adjacency matrix}

\item{copulaParam}{Copula parameter(S)}

\item{e_it}{Population dataset}

\item{B}{Strain-specific outbreak effects}

\item{T.prob}{Transition probabilities}

\item{r}{Trend component}

\item{s}{Seasonal component}

\item{u}{spatial component}
}
\value{
Multi-type outbreaks data
}
\description{
Title
}
\examples{
mod0<- simulateMultiModel(Modeltype = 0, time = 60, nstrain = 5, adj.matrix = sim_adjmat)
}
