% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inference.R
\name{SMOOTHING_INFERENCE}
\alias{SMOOTHING_INFERENCE}
\title{Title}
\usage{
SMOOTHING_INFERENCE(
  y,
  e_it,
  Modeltype,
  adjmat,
  step_sizes = list(r = 0.3, s = 0.3, u = 0.3),
  Stan = FALSE,
  MCMC_iterations = 15000,
  HMC_iterations = 5000,
  sdBs = 0.003,
  sdGs = 0.005,
  sdLambdas = 0.003,
  sdCop = 2e-04,
  y_total = NULL,
  RM_Gs = TRUE,
  RM_Bs = TRUE,
  RM_Cop = TRUE,
  RM_Lambdas = TRUE,
  burn_in = 30000,
  n_chains = 4,
  GPU = FALSE,
  adaptdelta = 0.9,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
